requirements.txt
aniso8601==9.0.1
astroid==3.0.2
asttokens==2.4.1
attrs==23.1.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
dill==0.3.7
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.19.0
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-Testing==0.8.1
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
ipython==8.18.1
isort==5.13.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_core==5.5.1
kaleido==0.2.1
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mccabe==0.7.0
nbformat==5.9.2
numpy==1.26.1
packaging==23.2
pandas==2.1.2
parso==0.8.3
pexpect==4.9.0
platformdirs==4.1.0
plotly==5.18.0
pluggy==1.3.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pylint==3.0.3
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.32.0
requests==2.31.0
rpds-py==0.15.2
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
stack-data==0.6.3
tenacity==8.2.3
threadpoolctl==3.2.0
tomli==2.0.1
tomlkit==0.12.3
traitlets==5.14.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
zipp==3.17.0
